requirements.txt
psycopg2-binary
uvicorn
fastapi
python-dotenv
python-multipart